+++ ouds_theme_orange_compact/lib/ouds_theme_orange_compact.dart
@@ -47,6 +47,7 @@ import 'package:ouds_theme_orange_compact/components/orangeCompact_switch_tokens
 import 'package:ouds_theme_orange_compact/components/orangeCompact_tag_tokens.dart';
 import 'package:ouds_theme_orange_compact/components/orangeCompact_textInput_tokens.dart';
 import 'package:ouds_theme_orange_compact/material/orangeCompact_material_color_tokens.dart';
+import 'package:ouds_theme_orange_compact/orange_compact_typography.dart';
 import 'package:ouds_theme_orange_compact/semantic/orangeCompact_border_semantic_tokens.dart';
 import 'package:ouds_theme_orange_compact/semantic/orangeCompact_color_semantic_tokens.dart';
 import 'package:ouds_theme_orange_compact/semantic/orangeCompact_elevation_semantic_tokens.dart';
@@ -61,6 +62,79 @@ import 'package:ouds_theme_orange_compact/semantic/orangeCompact_space_semantic_
 /// and component tokens. The class handles the theme's data for both light and dark modes, allowing dynamic
 /// switching based on the current [ThemeMode] setting.
 ///
+/// This theme uses the Helvetica Neue font family. **Due to legal issues Helvetica Neue font files are not bundled with this library.**
+///
+/// The Helvetica Neue font files for the Orange theme are available at
+/// [https://brand.orange.com/en/brand-basics/typography](https://brand.orange.com/en/brand-basics/typography) and can be used by copying the `ttf` files
+/// in the `assets/fonts` directory of the project
+///
+/// Load Fonts Before App Starts
+///
+/// **Usage example:**
+///
+/// ```dart
+///   // This example loads fonts from local assets.
+///   final orangeFontFamily = await OrangeFontFamily.getFontFamily(
+///     fontConfigs: [
+///       OrangeFontFamily(
+///         familyName: "Helvetica Neue Arabic",
+///         asset: "assets/fonts/helvetica_neue_arabic.ttf",
+///         fontWeight: FontWeight.w400,
+///         script: OrangeFontScript.arabic,
+///       ),
+///       OrangeFontFamily(
+///         familyName: "Helvetica Neue Arabic Bold",
+///         asset: "assets/fonts/helvetica_neue_arabic_bold.ttf",
+///         fontWeight: FontWeight.w700,
+///         script: OrangeFontScript.arabic,
+///       ),
+///       OrangeFontFamily(
+///           familyName: "Helvetica Neue Regular",
+///           asset: "assets/fonts/helvetica_neue_latin_roman.ttf",
+///           fontWeight: FontWeight.w400,
+///           script: OrangeFontScript.latin,
+///       ),
+///       OrangeFontFamily(
+///           familyName: "Helvetica Neue Medium",
+///           asset: "assets/fonts/helvetica_neue_latin_medium.ttf",
+///           fontWeight: FontWeight.w500,
+///           script: OrangeFontScript.latin,
+///       )
+///     ],
+///   );
+///
+/// OrangeCompactTheme(fontFamily);
+/// }
+///```
+///
+/// Although the preferred way of using the Helvetica Neue font is configuring bundled font files,
+/// there are some cases where this is not possible, for instance in open source projects where the font
+/// files cannot be bundled due to legal issues. In these cases, the font files can alternatively,
+/// the font files can be downloaded using Flutter's Downloadable Fonts feature
+/// by calling the [OrangeFontService.getFontFamily] method in your `app` before the `runApp` method in your main file.
+///
+///
+/// ```dart
+/// Future<void> main() async {
+///   WidgetsFlutterBinding.ensureInitialized();
+///
+///   final orangeFontFamily = await OrangeFontFamily.getFontFamily();
+///
+///   runApp(OudsApplication(fontFamily:orangeFontFamily));
+/// }
+///```
+///
+/// When using OrangeCompactTheme, simply provide the font as a parameter.
+///
+/// ```dart
+///   OrangeCompactTheme(
+///     orangeFontFamily
+///   )
+///```
+///
+/// Please note that the Flutter Downloadable Font feature works asynchronously, whether the font is already downloaded or not,
+/// and that default font family will be used if download fails.
+///
 /// The theme includes the following:
 /// - A color scheme with semantic tokens for light and dark modes,
 /// - Typography settings based on the defined font tokens,
@@ -69,7 +143,8 @@ import 'package:ouds_theme_orange_compact/semantic/orangeCompact_space_semantic_
 /// This theme is used for the visual styling of the Flutter application, ensuring a consistent user experience
 /// across both Android and iOS platforms.
 class OrangeCompactTheme implements OudsThemeContract {
-  OrangeCompactTheme();
+  String? orangeFontFamily;
+  OrangeCompactTheme(this.orangeFontFamily);
 
   @override
   String get name => "Orange Compact";
@@ -79,8 +154,6 @@ class OrangeCompactTheme implements OudsThemeContract {
 
   @override
   ThemeData get themeData => ThemeData(
-        fontFamily: fontFamily,
-        package: packageName,
         colorScheme: ColorScheme(
           brightness: Brightness.light,
           primary: materialColorTokens.primaryLight,
@@ -133,6 +206,7 @@ class OrangeCompactTheme implements OudsThemeContract {
             fontSize: fontTokens.sizeBodyMediumMobile,
             fontWeight: fontTokens.weightBodyDefault,
             color: colorSemanticTokens.contentColorTokens.contentDefaultLight,
+            fontFamily: fontFamily,
           ),
         ),
       );
@@ -162,7 +236,7 @@ class OrangeCompactTheme implements OudsThemeContract {
   OudsFontSemanticTokens get fontTokens => OrangeCompactFontSemanticTokens();
 
   @override
-  String get fontFamily => defaultTargetPlatform == TargetPlatform.android ? 'Roboto' : 'SFProDisplay';
+  String get fontFamily => orangeFontFamily ?? (defaultTargetPlatform == TargetPlatform.android ? 'Roboto' : 'SFProDisplay');
 
   @override
   String get packageName => 'ouds_theme_orange_compact';
@@ -188,7 +262,7 @@ class OrangeCompactTheme implements OudsThemeContract {
   }
 
   @override
-  OudsTypography get typographyTokens => OudsTypography();
+  OudsTypography get typographyTokens => OrangeCompactTypography();
 
   @override
   OudsProvidersTokens providersTokens(BuildContext context) {
@@ -264,8 +338,6 @@ class OrangeCompactTheme implements OudsThemeContract {
   @override
   ThemeData get darkThemeData {
     return ThemeData(
-      fontFamily: fontFamily,
-      package: packageName,
       colorScheme: ColorScheme(
         brightness: Brightness.dark,
         primary: materialColorTokens.primaryDark,
@@ -318,6 +390,7 @@ class OrangeCompactTheme implements OudsThemeContract {
           fontSize: fontTokens.sizeBodyMediumMobile,
           fontWeight: fontTokens.weightBodyDefault,
           color: colorSemanticTokens.contentColorTokens.contentDefaultDark,
+          fontFamily: fontFamily,
         ),
       ),
     );